+++ l10n/si.js
@@ -15,6 +15,7 @@ OC.L10N.register(
     "Select folder" : "බහාලුම තෝරන්න",
     "Cancel" : "අවලංගු කරන්න",
     "Files" : "ගොනු",
+    "Delete" : "Delete",
     "Install web app" : "වියමන යෙදුම ස්ථාපනය කරන්න",
     "Need help?" : "උපකාරයක් අවශ්‍යයද?",
     "Link" : "සබැඳිය",